week4/requirements.txt
mcp
langchain-core
langchain-anthropic
pandas
python-dotenv
